requirements.txt
flask
flask-restx
Flask-SQLAlchemy
SQLAlchemy
gunicorn
flask_cors
python-dotenv
pandas
